# Conflicts:
#	CarPooling/app/src/main/java/it/polito/mad/group27/carpooling/ui/trip/tripedit/TripEditFragment.kt
#	CarPooling/app/src/main/res/values-it/strings.xml
#	CarPooling/app/src/main/res/values/strings.xml

+++ CarPooling/app/src/main/java/it/polito/mad/group27/carpooling/ui/trip/Trip.kt
@@ -69,7 +69,7 @@ object BigDecimalSerializer: KSerializer<BigDecimal> {
 @Serializable
 @Parcelize
 data class Trip(
-    var id: Long = -1,
+    var id: Int = -1,
     @Serializable(with=UriSerializer::class)
     var carImageUri: Uri? = null,
     var totalSeats: Int? = null,
